🏗️ Split `feed` thunks into specific feature folders
In the new feature-based architecture, each feature has its own dedicated `store` subfolder. Thunks are the second part migrated into it.
