+++ src/features/activity-feed/post/components/ActivityFeed/index.jsx
@@ -3,7 +3,7 @@ import { useSelector, useDispatch } from 'react-redux'
 import { Box, Grid, CircularProgress } from '@mui/material'
 import { PostIdProvider } from '../../contexts/PostIdProvider'
 import { selectPosts, selectActivityFeedHasMorePosts, selectPostLoading } from '../../store/postsSelectors'
-import { fetchPosts } from '../../../../../redux/thunks/feed'
+import { fetchPosts } from '../../store/postsThunks'
 import { cleanFeedState } from '../../../../../redux/reducers/feed'
 import { ProfileUserIdContext } from '../../contexts/ProfileUserIdProvider'
 import { ActivityFeedHeader } from '../ActivityFeedHeader'
